requirements.txt
jupyter
langchain
langchain-community
langgraph
fastapi
uvicorn
python-dotenv
